adding functionality to users page
