+++ app/users/page.module.css
@@ -1,0 +1,186 @@
+.container {
+  display: flex;
+  flex-direction: row;
+  align-items: flex-start;
+  gap: 20px;
+  margin: 50px 125px 0;
+}
+
+.frame {
+  display: inline-flex;
+  background-color: #ffffff;
+  border-radius: 20px;
+  box-shadow: 0px 5px 11px #0000000f, 0px 21px 21px #0000000d, 0px 47px 28px #00000008, 0px 83px 33px #00000003,
+      0px 130px 37px transparent;
+}
+
+.stats .div {
+  display: flex;
+  flex-direction: row;
+  margin: 30px 50px;
+  gap: 60px;
+}
+
+.stats .column {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  gap: 10px;
+}
+
+.stats .title {
+  font-family: "Inter-SemiBold", Helvetica;
+  font-weight: 600;
+  color: #000000;
+  font-size: 14px;
+  letter-spacing: 0;
+  line-height: normal;
+  white-space: nowrap;
+}
+
+.stats .value {
+  font-family: "Inter-SemiBold", Helvetica;
+  font-weight: 600;
+  color: #000000;
+  font-size: 14px;
+  letter-spacing: 0;
+  line-height: normal;
+  white-space: nowrap;
+  text-align: center;
+}
+
+.jobs {
+  width: 100%;
+}
+
+.jobs .frame {
+  width: 100%;
+}
+
+.jobs .div {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  margin: 30px 40px;
+}
+
+.jobs .header .row {
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+}
+
+.jobs .cont .title {
+  font-family: "Inter-Bold", Helvetica;
+  font-weight: 700;
+  color: #737373;
+  font-size: 20px;
+  letter-spacing: 0;
+  line-height: normal;
+  margin-right: 12px;
+}
+
+.jobs .cont .titleUnderline {
+  display: flex;
+  height: 5px;
+  background-color: #737373;
+}
+
+.jobs .addBtn {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  gap: 10px;
+  padding: 5px 15px;
+  border-radius: 12px;
+  background-color: #1DBF73;
+  cursor: pointer;
+}
+
+.jobs .addBtn p {
+  font-family: "Inter-Medium", Helvetica;
+  font-weight: 700;
+  color: #ffffff;
+  font-size: 14px;
+  letter-spacing: 0;
+  line-height: normal;
+  white-space: nowrap;
+}
+
+.jobs .addBtn img {
+  filter: invert(100%) sepia(0%) saturate(7500%) hue-rotate(194deg) brightness(105%) contrast(100%);
+}
+
+.jobs .filters .row {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  flex-flow: wrap;
+  gap: 10px;
+  margin-top: 10px;
+}
+
+.jobs .filters .option {
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  padding: 2px 10px;
+  position: relative;
+  background-color: #d9d9d9;
+  border-radius: 7px;
+  overflow: hidden;
+  cursor: pointer;
+  font-family: "Inter-Regular", Helvetica;
+  font-weight: 400;
+  color: #000000;
+  font-size: 12px;
+  text-align: center;
+  letter-spacing: 0;
+  line-height: normal;
+}
+
+.jobs .filters .option:hover {
+  background-color: #bfbfbf;
+}
+
+.jobs .filters input {
+  display: none;
+}
+
+.jobs .filters .option:has(input:checked) {
+  background-color: #1DBF73;
+  color: #ffffff;
+}
+
+@media only screen and (max-width: 1440px) {
+  .container {
+    margin: 50px 50px 0;
+  }
+}
+
+/* @media only screen and (min-width: 950px) {
+  .jobsSM {
+    display: none;
+  }
+} */
+
+@media only screen and (max-width: 950px) {
+  .container {
+    margin: 50px 25px 0;
+    flex-direction: column;
+    align-items: center;
+  }
+}
+
+@media only screen and (max-width: 500px) {
+  .container {
+    margin: 50px 15px 0;
+  }
+
+  .jobs .div {
+    margin: 30px 20px;
+  }
+}
